backend/requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-multipart==0.0.22
slowapi==0.1.9
urllib3==2.6.3
argon2-cffi==25.1.0
sqlcipher3==0.6.2
pywin32==311; sys_platform == "win32"
sqlalchemy==2.0.44
pydantic==2.12.5
pydantic-settings==2.13.0
recipe-scrapers==15.11.0
beautifulsoup4==4.14.3
httpx==0.28.1
python-dateutil==2.9.0.post0
platformdirs==4.3.7
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
